// src/international-space-bar-server/openresponses/generated/zod/webSocketResponseCreateEventSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { WebSocketResponseCreateEvent } from "../types/WebSocketResponseCreateEvent.js";
import { createResponseBodySchema } from "./createResponseBodySchema.js";
import { z } from "zod/v4";

/**
 * @description A client-sent WebSocket event that starts a Responses turn. It uses the normal response-create body plus a required `type` field.
 */
export const webSocketResponseCreateEventSchema = z.lazy(() => createResponseBodySchema).and(z.object({
    "type": z.enum(["response.create"]).default("response.create").describe("The client event type. Always `response.create`.")
    })).describe("A client-sent WebSocket event that starts a Responses turn. It uses the normal response-create body plus a required `type` field.") as unknown as z.ZodType<WebSocketResponseCreateEvent>

export type WebSocketResponseCreateEventSchema = WebSocketResponseCreateEvent